Updated to work from a clean build OpenCV version and include fixes

+++ src/rgbd/depth_to_3d.h
@@ -77,7 +77,7 @@ convertDepthToFloat(const cv::Mat& depth, const cv::Mat& mask, float scale, cv::
         v_mat(n_points, 0) = v;
         T depth_i = depth.at<T>(v, u);
 
-        if (cvIsNaN(depth_i) || (depth_i == std::numeric_limits<T>::min()) || (depth_i == std::numeric_limits<T>::max()))
+        if (cvIsNaN((float)depth_i) || (depth_i == std::numeric_limits<T>::min()) || (depth_i == std::numeric_limits<T>::max()))
           z_mat(n_points, 0) = std::numeric_limits<float>::quiet_NaN();
         else
           z_mat(n_points, 0) = depth_i * scale;
@@ -107,7 +107,7 @@ convertDepthToFloat(const cv::Mat& depth, float scale, const cv::Mat &uv_mat, cv
   {
     T depth_i = depth.at < T > ((*uv_iter)[1], (*uv_iter)[0]);
 
-    if (cvIsNaN(depth_i) || (depth_i == std::numeric_limits < T > ::min())
+    if (cvIsNaN((float)depth_i) || (depth_i == std::numeric_limits < T > ::min())
         || (depth_i == std::numeric_limits < T > ::max()))
       *z_mat_iter = std::numeric_limits<float>::quiet_NaN();
     else
